Made all response entities with some minor logic

+++ src/main/java/com/movierp/server/controllers/CinemaController.java
@@ -3,7 +3,6 @@ package com.movierp.server.controllers;
 import com.movierp.server.entities.CinesEntity;
 import com.movierp.server.services.CinemaService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,37 +19,51 @@ public class CinemaController {
         this.cinemaService = cinemaService;
     }
 
-    @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<CinesEntity> create(CinesEntity cinema) {
-        this.cinemaService.save(cinema);
-        HttpHeaders responseHeaders = new HttpHeaders();
-        return new ResponseEntity<>(cinema, responseHeaders, HttpStatus.OK);
-    }
-
     @RequestMapping(method = RequestMethod.GET)
-    public Collection<CinesEntity> findAll() {
-        return (Collection<CinesEntity>) this.cinemaService.findAll();
+    public ResponseEntity<Collection<CinesEntity>> findAll() {
+        Collection<CinesEntity> cinemas = (Collection<CinesEntity>) this.cinemaService.findAll();
+
+        if (cinemas.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        return new ResponseEntity<>(cinemas, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public ResponseEntity findOne(@PathVariable("id") long id) {
-        try {
-            CinesEntity cinema = this.cinemaService.findOne(id);
-            return new ResponseEntity<>(cinema, HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<>(ex, HttpStatus.OK);
-        }
+    public ResponseEntity<CinesEntity> findOne(@PathVariable("id") int id) {
+        CinesEntity cinema = this.cinemaService.findOne(id);
+        return new ResponseEntity<>(cinema, HttpStatus.FOUND);
+    }
+
+    @RequestMapping(method = RequestMethod.POST)
+    public ResponseEntity<CinesEntity> create(@RequestBody CinesEntity cinema) {
+        this.cinemaService.save(cinema);
+        return new ResponseEntity<>(cinema, HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public ResponseEntity<CinesEntity> update(@PathVariable("id") CinesEntity cinema) {
+    public ResponseEntity<CinesEntity> update(@PathVariable("id") int id, @RequestBody CinesEntity cinema) {
+        CinesEntity currentCinema = this.cinemaService.findOne(id);
+
+        if (currentCinema == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         this.cinemaService.save(cinema);
-        HttpHeaders responseHeaders = new HttpHeaders();
-        return new ResponseEntity<>(cinema, responseHeaders, HttpStatus.OK);
+
+        return new ResponseEntity<>(cinema, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable("id") long id) {
+    public ResponseEntity<CinesEntity> delete(@PathVariable("id") int id) {
+        CinesEntity cinema = this.cinemaService.findOne(id);
+
+        if (cinema == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         this.cinemaService.delete(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
